@@ -1,68 +1,99 @@
-<!-- @author zhengjie -->
 <template>
   <div class="icon-body">
-    <el-input v-model="name" style="position: relative;" clearable placeholder="请输入图标名称" @clear="filterIcons" @input.native="filterIcons">
-      <i slot="suffix" class="el-icon-search el-input__icon" />
-    </el-input>
     <div class="icon-list">
-      <div v-for="(item, index) in iconList" :key="index" @click="selectedIcon(item)">
-        <svg-icon :icon-class="item" style="height: 30px;width: 16px;" />
-        <span>{{ item }}</span>
+      <div
+        v-for="(item, index) in nameArr"
+        :key="index"
+        @click="selectedIcon(item)"
+        :class="{ highlight: props.iconVal && item === props.iconVal }"
+      >
+        <svg-icon :icon-class="item" />
+        <span class="name">{{ item }}</span>
       </div>
     </div>
   </div>
 </template>
 
-<script>
-import icons from './requireIcons'
-export default {
-  name: 'IconSelect',
-  data() {
-    return {
-      name: '',
-      iconList: icons
-    }
+<script setup lang="ts">
+import nameArr from "./requireIcons";
+
+const emits = defineEmits(["selectIcon"]);
+
+const props = defineProps({
+  iconVal: {
+    type: String,
+    default: " ",
   },
-  methods: {
-    filterIcons() {
-      this.iconList = icons
-      if (this.name) {
-        this.iconList = this.iconList.filter(item => item.includes(this.name))
-      }
-    },
-    selectedIcon(name) {
-      this.$emit('selected', name)
-      document.body.click()
-    },
-    reset() {
-      this.name = ''
-      this.iconList = icons
-    }
-  }
-}
+});
+
+// 点击图标时触发
+const selectedIcon = (name: string) => {
+  emits("selectIcon", name);
+};
 </script>
 
-<style rel="stylesheet/scss" lang="scss" scoped>
-  .icon-body {
-    width: 100%;
-    padding: 10px;
-    .icon-list {
-      height: 200px;
-      overflow-y: scroll;
+<style lang="scss" scoped>
+.icon-body {
+  width: 100%;
+  padding: 10px;
+
+  .icon-list {
+    height: 200px;
+    overflow-y: scroll;
+    display: flex;
+    flex-wrap: wrap; // 使用 flex 布局，自动适应移动设备
+    justify-content: space-between;
+
+    @media (min-width: 1400px) {
       div {
-        height: 30px;
-        line-height: 30px;
-        margin-bottom: -5px;
-        cursor: pointer;
-        width: 33%;
-        float: left;
-      }
-      span {
-        display: inline-block;
-        vertical-align: -0.15em;
-        fill: currentColor;
-        overflow: hidden;
+        width: 25%; // 大屏幕：每行显示 4 个图标
       }
     }
+
+    @media (min-width: 768px) and (max-width: 1399px) {
+      div {
+        width: 25%; // 中屏幕：每行显示 4 个图标
+      }
+      .name {
+        display: none;
+      }
+    }
+
+    @media (max-width: 767px) {
+      div {
+        width: 33.33%; // 小屏幕：每行显示 3 个图标
+      }
+      .name {
+        display: none;
+      }
+    }
+
+    div {
+      height: 50px; // 设置高度以便于居中
+      margin-bottom: 10px;
+      cursor: pointer;
+      transition: background-color 0.3s;
+      display: flex; // 设置为 flex 布局
+      align-items: center; // 垂直居中
+      justify-content: center; // 水平居中
+    }
+
+    .svg-icon {
+      height: 30px;
+      margin-right: 5px; // 图标和文本之间的间距
+    }
+
+    span {
+      display: inline-block;
+      vertical-align: middle; // 中垂直对齐
+      fill: currentColor;
+      overflow: hidden;
+    }
   }
+
+  .highlight {
+    background-color: #e0f7fa;
+    border-radius: 5px;
+  }
+}
 </style>
