@@ -1,58 +1,64 @@
 <template>
   <div>
-    <svg-icon :icon-class="isFullscreen?'exit-fullscreen':'fullscreen'" @click="click" />
+    <svg-icon :icon-class="isFullscreen ? 'exit-fullscreen' : 'fullscreen'" @click="click" />
   </div>
 </template>
 
 <script>
-import screenfull from 'screenfull'
+import { defineComponent } from 'vue';
+import screenfull, { bindF11, unbindF11 } from '@/utils/screenfull';
 
-export default {
+export default defineComponent({
   name: 'Screenfull',
   data() {
     return {
       isFullscreen: false
-    }
+    };
   },
   mounted() {
-    this.init()
+    this.init();
+    bindF11();
   },
-  beforeDestroy() {
-    this.destroy()
+  beforeUnmount() {
+    this.destroy();
+    unbindF11();
   },
   methods: {
     click() {
-      if (!screenfull.enabled) {
-        this.$message({
+      console.log('screenfull.isEnabled=', screenfull.isEnabled, 'screenfull.isFullscreen=', screenfull.isFullscreen);
+      if (!screenfull.isEnabled) {
+        ElMessage({
           message: 'you browser can not work',
           type: 'warning'
-        })
-        return false
+        });
+        return false;
       }
-      screenfull.toggle()
+      screenfull.toggle();
     },
-    change() {
-      this.isFullscreen = screenfull.isFullscreen
+    handleChange() {
+      console.log('screenfull.handleChange', screenfull.isFullscreen);
+      this.isFullscreen = screenfull.isFullscreen;
     },
     init() {
-      if (screenfull.enabled) {
-        screenfull.on('change', this.change)
+      if (screenfull.isEnabled) {
+        screenfull.on('change', this.handleChange);
       }
     },
     destroy() {
-      if (screenfull.enabled) {
-        screenfull.off('change', this.change)
+      if (screenfull.isEnabled) {
+        screenfull.off('change', this.handleChange);
       }
     }
   }
-}
+});
 </script>
 
 <style scoped>
 .screenfull-svg {
   display: inline-block;
   cursor: pointer;
-  fill: #5a5e66;;
+  fill: #5a5e66;
+  ;
   width: 20px;
   height: 20px;
   vertical-align: 10px;
