@@ -1,45 +1,46 @@
-import store from '@/store'
+import store from '@/store';
+import { defineComponent } from 'vue';
 
-const { body } = document
-const WIDTH = 992 // refer to Bootstrap's responsive design
+const { body } = document;
+const WIDTH = 992; // refer to Bootstrap's responsive design
 
-export default {
+export default defineComponent({
   watch: {
-    $route(route) {
+    $route() {
       if (this.device === 'mobile' && this.sidebar.opened) {
-        store.dispatch('app/closeSideBar', { withoutAnimation: false })
+        store.app().closeSidebar({ withoutAnimation: false });
       }
     }
   },
   beforeMount() {
-    window.addEventListener('resize', this.$_resizeHandler)
+    window.addEventListener('resize', this.resizeHandler);
   },
-  beforeDestroy() {
-    window.removeEventListener('resize', this.$_resizeHandler)
+  beforeUnmount() {
+    window.removeEventListener('resize', this.resizeHandler);
   },
   mounted() {
-    const isMobile = this.$_isMobile()
+    const isMobile = this.isMobile();
     if (isMobile) {
-      store.dispatch('app/toggleDevice', 'mobile')
-      store.dispatch('app/closeSideBar', { withoutAnimation: true })
+      store.app().toggleDevice('mobile');
+      store.app().closeSidebar({ withoutAnimation: true });
     }
   },
   methods: {
-    // use $_ for mixins properties
+    // do not use $_ for mixins properties
     // https://vuejs.org/v2/style-guide/index.html#Private-property-names-essential
-    $_isMobile() {
-      const rect = body.getBoundingClientRect()
-      return rect.width - 1 < WIDTH
+    isMobile() {
+      const rect = body.getBoundingClientRect();
+      return rect.width - 1 < WIDTH;
     },
-    $_resizeHandler() {
+    resizeHandler() {
       if (!document.hidden) {
-        const isMobile = this.$_isMobile()
-        store.dispatch('app/toggleDevice', isMobile ? 'mobile' : 'desktop')
+        const isMobile = this.isMobile();
+        store.app().toggleDevice(isMobile ? 'mobile' : 'desktop');
 
         if (isMobile) {
-          store.dispatch('app/closeSideBar', { withoutAnimation: true })
+          store.app().closeSidebar({ withoutAnimation: true });
         }
       }
     }
   }
-}
+});
