@@ -1,9 +1,15 @@
 <template>
-  <div class="login-container" :style="{backgroundImage:`url(${imgSrc})`}">
-    <el-form ref="loginForm" :model="loginForm" :rules="loginRules" class="login-form" autocomplete="on" label-position="left">
-
+  <div class="login-container">
+    <el-form
+      ref="loginForm"
+      :model="loginForm"
+      :rules="loginRules"
+      class="login-form"
+      autocomplete="on"
+      label-position="left"
+    >
       <div class="title-container">
-        <h3 class="title">Go Web Mini</h3>
+        <h3 class="title">登录</h3>
       </div>
 
       <el-form-item prop="username">
@@ -13,7 +19,7 @@
         <el-input
           ref="username"
           v-model="loginForm.username"
-          placeholder="用户名"
+          placeholder="Username"
           name="username"
           type="text"
           tabindex="1"
@@ -21,7 +27,12 @@
         />
       </el-form-item>
 
-      <el-tooltip v-model="capsTooltip" content="Caps lock is On" placement="right" manual>
+      <el-tooltip
+        v-model="capsTooltip"
+        content="Caps lock is On"
+        placement="right"
+        manual
+      >
         <el-form-item prop="password">
           <span class="svg-container">
             <svg-icon icon-class="password" />
@@ -31,147 +42,236 @@
             ref="password"
             v-model="loginForm.password"
             :type="passwordType"
-            placeholder="密码"
+            placeholder="Password"
             name="password"
             tabindex="2"
             autocomplete="on"
-            @keyup.native="checkCapslock"
+            @keyup="checkCapslock"
             @blur="capsTooltip = false"
-            @keyup.enter.native="handleLogin"
+            @keyup.enter="handleLogin"
           />
           <span class="show-pwd" @click="showPwd">
-            <svg-icon :icon-class="passwordType === 'password' ? 'eye' : 'eye-open'" />
+            <svg-icon
+              :icon-class="passwordType === 'password' ? 'eye' : 'eye-open'"
+            />
           </span>
         </el-form-item>
       </el-tooltip>
 
-      <el-button :loading="loading" type="primary" class="login-btn" @click.native.prevent="handleLogin">Login</el-button>
+      <el-button
+        :loading="loading"
+        type="primary"
+        style="width: 100%; margin-bottom: 30px"
+        @click.prevent="handleLogin"
+      >
+        Login</el-button
+      >
 
+      <div style="position: relative">
+        <div class="tips">
+          <span>Username : admin</span>
+          <span>Password : any</span>
+        </div>
+        <div class="tips">
+          <span style="margin-right: 18px">Username : editor</span>
+          <span>Password : any</span>
+        </div>
+
+        <el-button
+          class="thirdparty-button"
+          type="primary"
+          @click="showDialog = true"
+        >
+          Or connect with
+        </el-button>
+      </div>
     </el-form>
 
+    <el-dialog title="Or connect with" v-model="showDialog">
+      Can not be simulated on local, so please combine you own business
+      simulation! ! !
+      <br />
+      <br />
+      <br />
+      <social-sign />
+    </el-dialog>
   </div>
 </template>
 
-<script>
-import JSEncrypt from 'jsencrypt'
+<script lang="ts">
+import { validUsername } from "@/utils/validate";
+import { defineComponent } from "vue";
+import SocialSign from "./components/SocialSignin.vue";
+import type { FormItemRule } from "element-plus";
+import type { IForm } from "@/types/element-plus";
+import store from "@/store";
 
-export default {
-  name: 'Login',
+interface QueryType {
+  // 自定义key 任意字符串
+  [propname: string]: string;
+}
+
+export default defineComponent({
+  name: "Login",
+  components: { SocialSign },
   data() {
-    const validatePassword = (rule, value, callback) => {
-      if (value.length < 6) {
-        callback(new Error('The password can not be less than 6 digits'))
+    const validateUsername: FormItemRule["validator"] = (
+      _rule,
+      value,
+      callback
+    ) => {
+      if (!validUsername(value)) {
+        callback(new Error("Please enter the correct user name"));
       } else {
-        callback()
+        callback();
       }
-    }
+    };
+    const validatePassword: FormItemRule["validator"] = (
+      _rule,
+      value,
+      callback
+    ) => {
+      if (value.length < 6) {
+        callback(new Error("The password can not be less than 6 digits"));
+      } else {
+        callback();
+      }
+    };
     return {
-      imgSrc: require('@/assets/backgd-image/backimg.jpg'),
       loginForm: {
-        username: '',
-        password: ''
+        username: "admin",
+        password: "123456",
       },
       loginRules: {
-        username: [{ required: true, trigger: 'blur' }],
-        password: [{ required: true, trigger: 'blur', validator: validatePassword }]
+        username: [
+          { required: true, message: "Username is required", trigger: "blur" },
+          {
+            min: 3,
+            message: "Username must be at least 3 characters",
+            trigger: "blur",
+          },
+        ],
+        password: [
+          { required: true, message: "Password is required", trigger: "blur" },
+          {
+            min: 6,
+            message: "Password must be at least 6 characters",
+            trigger: "blur",
+          },
+        ],
       },
-      passwordType: 'password',
-      publicKey: `-----BEGIN PUBLIC KEY-----
-MIGfMA0GCSqGSIb3DQEBAQUAA4GNADCBiQKBgQDbOYcY8HbDaNM9ooYXoc9s+R5o
-R05ZL1BsVKadQBgOVH/kj7PQuD+ABEFVgB6rJNi287fRuZeZR+MCoG72H+AYsAhR
-sEaB5SuI7gDEstXuTyjhx5bz0wUujbDK4VMgRfPO6MQo+A0c95OadDEvEQDG3KBQ
-wLXapv+ZfsjG7NgdawIDAQAB
------END PUBLIC KEY-----`,
+
+      passwordType: "password",
       capsTooltip: false,
       loading: false,
+      showDialog: false,
       redirect: undefined,
-      otherQuery: {}
-    }
+      otherQuery: {},
+    };
   },
   watch: {
     $route: {
-      handler: function(route) {
-        const query = route.query
+      handler: function (route) {
+        const query = route.query;
         if (query) {
-          this.redirect = query.redirect
-          this.otherQuery = this.getOtherQuery(query)
+          this.redirect = query.redirect;
+          this.otherQuery = this.getOtherQuery(query);
         }
       },
-      immediate: true
-    }
+      immediate: true,
+    },
   },
   created() {
     // window.addEventListener('storage', this.afterQRScan)
   },
   mounted() {
-    if (this.loginForm.username === '') {
-      this.$refs.username.focus()
-    } else if (this.loginForm.password === '') {
-      this.$refs.password.focus()
+    if (this.loginForm.username === "") {
+      (this.$refs.username as HTMLElement).focus();
+    } else if (this.loginForm.password === "") {
+      (this.$refs.password as HTMLElement).focus();
     }
   },
-  destroyed() {
+  unmounted() {
     // window.removeEventListener('storage', this.afterQRScan)
   },
   methods: {
     checkCapslock(e) {
-      const { key } = e
-      this.capsTooltip = key && key.length === 1 && (key >= 'A' && key <= 'Z')
+      const { key } = e;
+      this.capsTooltip = key && key.length === 1 && key >= "A" && key <= "Z";
     },
     showPwd() {
-      if (this.passwordType === 'password') {
-        this.passwordType = ''
+      if (this.passwordType === "password") {
+        this.passwordType = "";
       } else {
-        this.passwordType = 'password'
+        this.passwordType = "password";
       }
       this.$nextTick(() => {
-        this.$refs.password.focus()
-      })
+        (this.$refs.password as HTMLElement).focus();
+      });
     },
     handleLogin() {
-      this.$refs.loginForm.validate(valid => {
-        if (valid) {
-          this.loading = true
-          // 密码RSA加密处理
-          const encryptor = new JSEncrypt()
-          // 设置公钥
-          encryptor.setPublicKey(this.publicKey)
-          // 加密密码
-         // const encPassword = encryptor.encrypt(this.loginForm.password)
-          const encPassword = this.loginForm.password
-          const encLoginForm = { username: this.loginForm.username, password: encPassword }
-          this.$store.dispatch('user/login', encLoginForm)
-            .then(() => {
-              this.$router.push({ path: this.redirect || '/', query: this.otherQuery })
-              this.loading = false
-            })
-            .catch(() => {
-              this.loading = false
-            })
-        } else {
-          console.log('error submit!!')
-          return false
-        }
-      })
+      (this.$refs.loginForm as IForm).validate((valid) => {
+        return new Promise((resolve, reject) => {
+          if (valid) {
+            this.loading = true;
+            store
+              .user()
+              .login(this.loginForm)
+              .then(() => {
+                this.$router.push({
+                  path: this.redirect || "/",
+                  query: this.otherQuery,
+                });
+                this.loading = false;
+              })
+              .catch(() => {
+                this.loading = false;
+              })
+              .finally(() => {
+                resolve();
+              });
+          } else {
+            reject();
+          }
+        });
+      });
     },
-    getOtherQuery(query) {
-      return Object.keys(query).reduce((acc, cur) => {
-        if (cur !== 'redirect') {
-          acc[cur] = query[cur]
+    getOtherQuery(query: QueryType) {
+      return Object.keys(query).reduce((acc: QueryType, cur) => {
+        if (cur !== "redirect") {
+          acc[cur] = query[cur];
         }
-        return acc
-      }, {})
-    }
-  }
-}
+        return acc;
+      }, {});
+    },
+    // afterQRScan() {
+    //   if (e.key === 'x-admin-oauth-code') {
+    //     const code = getQueryObject(e.newValue)
+    //     const codeMap = {
+    //       wechat: 'code',
+    //       tencent: 'code'
+    //     }
+    //     const type = codeMap[this.auth_type]
+    //     const codeName = code[type]
+    //     if (codeName) {
+    //       store.user().LoginByThirdparty(codeName).then(() => {
+    //         this.$router.push({ path: this.redirect || '/' })
+    //       })
+    //     } else {
+    //       alert('第三方登录失败')
+    //     }
+    //   }
+    // }
+  },
+});
 </script>
 
 <style lang="scss">
 /* 修复input 背景不协调 和光标变色 */
 /* Detail see https://github.com/PanJiaChen/vue-element-admin/pull/927 */
 
-$bg:#5c646d;
-$light_gray:#fff;
+$bg: #283443;
+$light_gray: #fff;
 $cursor: #fff;
 
 @supports (-webkit-mask: none) and (not (cater-color: $cursor)) {
@@ -180,18 +280,22 @@ $cursor: #fff;
   }
 }
 
-/* reset element-ui css */
+/* reset element-plus css */
 .login-container {
   .el-input {
-    display: inline-block;
     height: 47px;
     width: 85%;
 
+    .el-input__wrapper,
     input {
       background: transparent;
       border: 0px;
       -webkit-appearance: none;
       border-radius: 0px;
+      box-shadow: none;
+    }
+
+    input {
       padding: 12px 5px 12px 15px;
       color: $light_gray;
       height: 47px;
@@ -206,59 +310,43 @@ $cursor: #fff;
 
   .el-form-item {
     border: 1px solid rgba(255, 255, 255, 0.1);
-    background: $bg;
+    background: rgba(0, 0, 0, 0.1);
     border-radius: 5px;
     color: #454545;
-    margin-left: 20px;
-    margin-right: 20px;
-  }
-
-  .login-btn{
-    width: calc(100% - 40px);
-    height: 45px;
-    margin-left: 20px;
-    margin-right: 20px;
-    margin-bottom: 30px;
-    border: 0;
-    font-size: 16px;
-    font-weight: bold;
-    // opacity: 0.9;
-    border-radius: 20px;
-    background-image: linear-gradient(to right, #74ebd5 0%, #9face6 100%);
-  }
-
-  .login-btn:hover{
-    color: greenyellow;
-    background-image: linear-gradient(to right, #eea2a2 0%, #bbc1bf 19%, #57c6e1 42%, #b49fda 79%, #7ac5d8 100%);
   }
 }
 </style>
 
 <style lang="scss" scoped>
-$bg:#5c646d;
-$dark_gray:#889aa4;
-$light_gray:#eee;
+$bg: #2d3a4b;
+$dark_gray: #889aa4;
+$light_gray: #eee;
 
 .login-container {
   min-height: 100%;
   width: 100%;
-  height: 100%;
   background-color: $bg;
-  // overflow: hidden;
-  background-size: cover;
-  display: flex;
-  justify-content: center;
-  align-items: center;
+  overflow: hidden;
 
   .login-form {
-    // position: relative;
+    position: relative;
     width: 520px;
     max-width: 100%;
-    // padding: 160px 35px 0;
-    // margin: 0 auto;
-    background-color: rgba(0,0,0,0.5);
-    // overflow: hidden;
-    border-radius: 8px;
+    padding: 160px 35px 0;
+    margin: 0 auto;
+    overflow: hidden;
+  }
+
+  .tips {
+    font-size: 14px;
+    color: #fff;
+    margin-bottom: 10px;
+
+    span {
+      &:first-of-type {
+        margin-right: 16px;
+      }
+    }
   }
 
   .svg-container {
@@ -275,10 +363,9 @@ $light_gray:#eee;
     .title {
       font-size: 26px;
       color: $light_gray;
-      margin: 20px auto 20px auto;
+      margin: 0px auto 40px auto;
       text-align: center;
       font-weight: bold;
-      // letter-spacing: 2px;
     }
   }
 
