@@ -6,14 +6,14 @@
  * @return {[blob]}      [description]
  */
 export default function(data, mime) {
-  data = data.split(',')[1]
-  data = window.atob(data)
-  var ia = new Uint8Array(data.length)
+  data = data.split(',')[1];
+  data = window.atob(data);
+  var ia = new Uint8Array(data.length);
   for (var i = 0; i < data.length; i++) {
-    ia[i] = data.charCodeAt(i)
+    ia[i] = data.charCodeAt(i);
   }
   // canvas.toDataURL 返回的默认格式就是 image/png
   return new Blob([ia], {
     type: mime
-  })
+  });
 }
