@@ -1,8 +1,8 @@
 <template>
-  <div :class="{active:isActive}" class="share-dropdown-menu">
+  <div :class="{ active: isActive }" class="share-dropdown-menu">
     <div class="share-dropdown-menu-wrapper">
       <span class="share-dropdown-menu-title" @click.self="clickTitle">{{ title }}</span>
-      <div v-for="(item,index) of items" :key="index" class="share-dropdown-menu-item">
+      <div v-for="(item, index) of items" :key="index" class="share-dropdown-menu-item">
         <a v-if="item.href" :href="item.href" target="_blank">{{ item.title }}</a>
         <span v-else>{{ item.title }}</span>
       </div>
@@ -11,13 +11,13 @@
 </template>
 
 <script>
-export default {
+import { defineComponent } from 'vue';
+
+export default defineComponent({
   props: {
     items: {
       type: Array,
-      default: function() {
-        return []
-      }
+      default: () => []
     },
     title: {
       type: String,
@@ -27,24 +27,26 @@ export default {
   data() {
     return {
       isActive: false
-    }
+    };
   },
   methods: {
     clickTitle() {
-      this.isActive = !this.isActive
+      this.isActive = !this.isActive;
     }
   }
-}
+});
 </script>
 
-<style lang="scss" >
+<style lang="scss">
 $n: 9; //和items.length 相同
 $t: .1s;
+
 .share-dropdown-menu {
   width: 250px;
   position: relative;
   z-index: 1;
-  height: auto!important;
+  height: auto !important;
+
   &-title {
     width: 100%;
     display: block;
@@ -56,11 +58,13 @@ $t: .1s;
     font-size: 20px;
     text-align: center;
     z-index: 2;
-    transform: translate3d(0,0,0);
+    transform: translate3d(0, 0, 0);
   }
+
   &-wrapper {
     position: relative;
   }
+
   &-item {
     text-align: center;
     position: absolute;
@@ -74,10 +78,12 @@ $t: .1s;
     overflow: hidden;
     opacity: 1;
     transition: transform 0.28s ease;
+
     &:hover {
       background: black;
       color: white;
     }
+
     @for $i from 1 through $n {
       &:nth-of-type(#{$i}) {
         z-index: -1;
@@ -86,10 +92,12 @@ $t: .1s;
       }
     }
   }
+
   &.active {
     .share-dropdown-menu-wrapper {
       z-index: 1;
     }
+
     .share-dropdown-menu-item {
       @for $i from 1 through $n {
         &:nth-of-type(#{$i}) {
