@@ -7,7 +7,9 @@
 </template>
 
 <script>
-export default {
+import { defineComponent } from 'vue';
+
+export default defineComponent({
   props: {
     className: {
       type: String,
@@ -18,7 +20,7 @@ export default {
       default: 'vue-element-admin'
     }
   }
-}
+});
 </script>
 
 <style>
