@@ -1,9 +1,9 @@
 <template>
   <div class="upload-container">
-    <el-button :style="{background:color,borderColor:color}" icon="el-icon-upload" size="mini" type="primary" @click=" dialogVisible=true">
+    <el-button :style="{background:color,borderColor:color}" :icon="IconUpload" size="small" type="primary" @click=" dialogVisible=true">
       upload
     </el-button>
-    <el-dialog :visible.sync="dialogVisible">
+    <el-dialog v-model="dialogVisible">
       <el-upload
         :multiple="true"
         :file-list="fileList"
@@ -30,9 +30,11 @@
 </template>
 
 <script>
+import { defineComponent, markRaw } from 'vue';
+import { Upload as IconUpload } from '@element-plus/icons-vue';
 // import { getToken } from 'api/qiniu'
 
-export default {
+export default defineComponent({
   name: 'EditorSlideUpload',
   props: {
     color: {
@@ -42,69 +44,73 @@ export default {
   },
   data() {
     return {
+      IconUpload: markRaw(IconUpload),
       dialogVisible: false,
       listObj: {},
       fileList: []
-    }
+    };
   },
   methods: {
     checkAllSuccess() {
-      return Object.keys(this.listObj).every(item => this.listObj[item].hasSuccess)
+      return Object.keys(this.listObj).every(item => this.listObj[item].hasSuccess);
     },
     handleSubmit() {
-      const arr = Object.keys(this.listObj).map(v => this.listObj[v])
+      const arr = Object.keys(this.listObj).map(v => this.listObj[v]);
       if (!this.checkAllSuccess()) {
-        this.$message('Please wait for all images to be uploaded successfully. If there is a network problem, please refresh the page and upload again!')
-        return
+        ElMessage({
+          message: 'Please wait for all images to be uploaded successfully. If there is a network problem, please refresh the page and upload again!'
+        });
+        return;
       }
-      this.$emit('successCBK', arr)
-      this.listObj = {}
-      this.fileList = []
-      this.dialogVisible = false
+      this.$emit('successCBK', arr);
+      this.listObj = {};
+      this.fileList = [];
+      this.dialogVisible = false;
     },
     handleSuccess(response, file) {
-      const uid = file.uid
-      const objKeyArr = Object.keys(this.listObj)
+      const uid = file.uid;
+      const objKeyArr = Object.keys(this.listObj);
       for (let i = 0, len = objKeyArr.length; i < len; i++) {
         if (this.listObj[objKeyArr[i]].uid === uid) {
-          this.listObj[objKeyArr[i]].url = response.files.file
-          this.listObj[objKeyArr[i]].hasSuccess = true
-          return
+          this.listObj[objKeyArr[i]].url = response.files.file;
+          this.listObj[objKeyArr[i]].hasSuccess = true;
+          return;
         }
       }
     },
     handleRemove(file) {
-      const uid = file.uid
-      const objKeyArr = Object.keys(this.listObj)
+      const uid = file.uid;
+      const objKeyArr = Object.keys(this.listObj);
       for (let i = 0, len = objKeyArr.length; i < len; i++) {
         if (this.listObj[objKeyArr[i]].uid === uid) {
-          delete this.listObj[objKeyArr[i]]
-          return
+          delete this.listObj[objKeyArr[i]];
+          return;
         }
       }
     },
     beforeUpload(file) {
-      const _self = this
-      const _URL = window.URL || window.webkitURL
-      const fileName = file.uid
-      this.listObj[fileName] = {}
-      return new Promise((resolve, reject) => {
-        const img = new Image()
-        img.src = _URL.createObjectURL(file)
+      // eslint-disable-next-line @typescript-eslint/no-this-alias
+      const _self = this;
+      const _URL = window.URL || window.webkitURL;
+      const fileName = file.uid;
+      this.listObj[fileName] = {};
+      return new Promise((resolve) => {
+        const img = new Image();
+        img.src = _URL.createObjectURL(file);
         img.onload = function() {
-          _self.listObj[fileName] = { hasSuccess: false, uid: file.uid, width: this.width, height: this.height }
-        }
-        resolve(true)
-      })
+          _self.listObj[fileName] = { hasSuccess: false, uid: file.uid, width: this.width, height: this.height };
+        };
+        resolve(true);
+      });
     }
   }
-}
+});
 </script>
 
 <style lang="scss" scoped>
 .editor-slide-upload {
   margin-bottom: 20px;
-  ::v-deep .el-upload--picture-card {
+  :deep(.el-upload--picture-card) {
     width: 100%;
   }
 }
