@@ -1,30 +1,31 @@
-let callbacks = []
+let callbacks = [];
 
 function loadedTinymce() {
   // to fixed https://github.com/PanJiaChen/vue-element-admin/issues/2144
   // check is successfully downloaded script
-  return window.tinymce
+  return window.tinymce;
 }
 
 const dynamicLoadScript = (src, callback) => {
-  const existingScript = document.getElementById(src)
-  const cb = callback || function() {}
+  const existingScript = document.getElementById(src);
+  // eslint-disable-next-line @typescript-eslint/no-empty-function
+  const cb = callback || function() {};
 
   if (!existingScript) {
-    const script = document.createElement('script')
-    script.src = src // src url for the third-party library being loaded.
-    script.id = src
-    document.body.appendChild(script)
-    callbacks.push(cb)
-    const onEnd = 'onload' in script ? stdOnEnd : ieOnEnd
-    onEnd(script)
+    const script = document.createElement('script');
+    script.src = src; // src url for the third-party library being loaded.
+    script.id = src;
+    document.body.appendChild(script);
+    callbacks.push(cb);
+    const onEnd = 'onload' in script ? stdOnEnd : ieOnEnd;
+    onEnd(script);
   }
 
   if (existingScript && cb) {
     if (loadedTinymce()) {
-      cb(null, existingScript)
+      cb(null, existingScript);
     } else {
-      callbacks.push(cb)
+      callbacks.push(cb);
     }
   }
 
@@ -32,28 +33,28 @@ const dynamicLoadScript = (src, callback) => {
     script.onload = function() {
       // this.onload = null here is necessary
       // because even IE9 works not like others
-      this.onerror = this.onload = null
+      this.onerror = this.onload = null;
       for (const cb of callbacks) {
-        cb(null, script)
+        cb(null, script);
       }
-      callbacks = null
-    }
+      callbacks = null;
+    };
     script.onerror = function() {
-      this.onerror = this.onload = null
-      cb(new Error('Failed to load ' + src), script)
-    }
+      this.onerror = this.onload = null;
+      cb(new Error('Failed to load ' + src), script);
+    };
   }
 
   function ieOnEnd(script) {
     script.onreadystatechange = function() {
-      if (this.readyState !== 'complete' && this.readyState !== 'loaded') return
-      this.onreadystatechange = null
+      if (this.readyState !== 'complete' && this.readyState !== 'loaded') return;
+      this.onreadystatechange = null;
       for (const cb of callbacks) {
-        cb(null, script) // there is no way to catch loading errors in IE8
+        cb(null, script); // there is no way to catch loading errors in IE8
       }
-      callbacks = null
-    }
+      callbacks = null;
+    };
   }
-}
+};
 
-export default dynamicLoadScript
+export default dynamicLoadScript;
