@@ -34,12 +34,13 @@
 </template>
 
 <script>
-import { getToken } from '@/api/qiniu'
+import { defineComponent } from 'vue';
+import { getToken } from '@/api/qiniu';
 
-export default {
+export default defineComponent({
   name: 'SingleImageUpload3',
   props: {
-    value: {
+    modelValue: {
       type: String,
       default: ''
     }
@@ -48,45 +49,46 @@ export default {
     return {
       tempUrl: '',
       dataObj: { token: '', key: '' }
-    }
+    };
   },
   computed: {
     imageUrl() {
-      return this.value
+      return this.modelValue;
     }
   },
   methods: {
     rmImage() {
-      this.emitInput('')
+      this.emitInput('');
     },
     emitInput(val) {
-      this.$emit('input', val)
+      this.$emit('update:modelValue', val);
     },
     handleImageSuccess(file) {
-      this.emitInput(file.files.file)
+      this.emitInput(file.files.file);
     },
     beforeUpload() {
-      const _self = this
+      // eslint-disable-next-line @typescript-eslint/no-this-alias
+      const _self = this;
       return new Promise((resolve, reject) => {
         getToken().then(response => {
-          const key = response.data.qiniu_key
-          const token = response.data.qiniu_token
-          _self._data.dataObj.token = token
-          _self._data.dataObj.key = key
-          this.tempUrl = response.data.qiniu_url
-          resolve(true)
+          const key = response.data.qiniu_key;
+          const token = response.data.qiniu_token;
+          _self._data.dataObj.token = token;
+          _self._data.dataObj.key = key;
+          this.tempUrl = response.data.qiniu_url;
+          resolve(true);
         }).catch(err => {
-          console.log(err)
-          reject(false)
-        })
-      })
+          console.log(err);
+          reject(false);
+        });
+      });
     }
   }
-}
+});
 </script>
 
 <style lang="scss" scoped>
-@import "~@/styles/mixin.scss";
+@import "@/styles/mixin.scss";
 .upload-container {
   width: 100%;
   position: relative;
