@@ -1,26 +1,29 @@
-export default {
+import store from '@/store';
+import { defineComponent } from 'vue';
+
+export default defineComponent({
   computed: {
     device() {
-      return this.$store.state.app.device
+      return store.app().device;
     }
   },
   mounted() {
     // In order to fix the click on menu on the ios device will trigger the mouseleave bug
     // https://github.com/PanJiaChen/vue-element-admin/issues/1135
-    this.fixBugIniOS()
+    this.fixBugIniOS();
   },
   methods: {
     fixBugIniOS() {
-      const $subMenu = this.$refs.subMenu
+      const $subMenu = this.$refs.subMenu;
       if ($subMenu) {
-        const handleMouseleave = $subMenu.handleMouseleave
+        const handleMouseleave = $subMenu.handleMouseleave;
         $subMenu.handleMouseleave = (e) => {
           if (this.device === 'mobile') {
-            return
+            return;
           }
-          handleMouseleave(e)
-        }
+          handleMouseleave(e);
+        };
       }
     }
   }
-}
+});
