@@ -1,10 +1,10 @@
-import defaultSettings from '@/settings'
+import defaultSettings from '@/settings';
 
-const title = defaultSettings.title || 'Vue Element Admin'
+const title = defaultSettings.title || '后台管理系统';
 
 export default function getPageTitle(pageTitle) {
   if (pageTitle) {
-    return `${pageTitle} - ${title}`
+    return `${pageTitle} - ${title}`;
   }
-  return `${title}`
+  return `${title}`;
 }
