@@ -1,4 +1,4 @@
-import store from '@/store'
+import store from '@/store';
 
 /**
  * @param {Array} value
@@ -7,15 +7,19 @@ import store from '@/store'
  */
 export default function checkPermission(value) {
   if (value && value instanceof Array && value.length > 0) {
-    const roles = store.getters && store.getters.roles
-    const permissionRoles = value
+    const roles = store.user && store.user().roles;
+    const permissionRoles = value;
 
     const hasPermission = roles.some(role => {
-      return permissionRoles.includes(role)
-    })
-    return hasPermission
+      return permissionRoles.includes(role);
+    });
+
+    if (!hasPermission) {
+      return false;
+    }
+    return true;
   } else {
-    console.error(`need roles! Like v-permission="['admin','editor']"`)
-    return false
+    console.error(`need roles! Like v-permission="['admin','editor']"`);
+    return false;
   }
 }
