@@ -1,29 +1,29 @@
 Math.easeInOutQuad = function(t, b, c, d) {
-  t /= d / 2
+  t /= d / 2;
   if (t < 1) {
-    return c / 2 * t * t + b
+    return c / 2 * t * t + b;
   }
-  t--
-  return -c / 2 * (t * (t - 2) - 1) + b
-}
+  t--;
+  return -c / 2 * (t * (t - 2) - 1) + b;
+};
 
 // requestAnimationFrame for Smart Animating http://goo.gl/sx5sts
 var requestAnimFrame = (function() {
-  return window.requestAnimationFrame || window.webkitRequestAnimationFrame || window.mozRequestAnimationFrame || function(callback) { window.setTimeout(callback, 1000 / 60) }
-})()
+  return window.requestAnimationFrame || window.webkitRequestAnimationFrame || window.mozRequestAnimationFrame || function(callback) { window.setTimeout(callback, 1000 / 60); };
+})();
 
 /**
  * Because it's so fucking difficult to detect the scrolling element, just move them all
  * @param {number} amount
  */
 function move(amount) {
-  document.documentElement.scrollTop = amount
-  document.body.parentNode.scrollTop = amount
-  document.body.scrollTop = amount
+  document.documentElement.scrollTop = amount;
+  document.body.parentNode.scrollTop = amount;
+  document.body.scrollTop = amount;
 }
 
 function position() {
-  return document.documentElement.scrollTop || document.body.parentNode.scrollTop || document.body.scrollTop
+  return document.documentElement.scrollTop || document.body.parentNode.scrollTop || document.body.scrollTop;
 }
 
 /**
@@ -32,27 +32,27 @@ function position() {
  * @param {Function} callback
  */
 export function scrollTo(to, duration, callback) {
-  const start = position()
-  const change = to - start
-  const increment = 20
-  let currentTime = 0
-  duration = (typeof (duration) === 'undefined') ? 500 : duration
+  const start = position();
+  const change = to - start;
+  const increment = 20;
+  let currentTime = 0;
+  duration = (typeof (duration) === 'undefined') ? 500 : duration;
   var animateScroll = function() {
     // increment the time
-    currentTime += increment
+    currentTime += increment;
     // find the value with the quadratic in-out easing function
-    var val = Math.easeInOutQuad(currentTime, start, change, duration)
+    var val = Math.easeInOutQuad(currentTime, start, change, duration);
     // move the document.body
-    move(val)
+    move(val);
     // do the animation unless its over
     if (currentTime < duration) {
-      requestAnimFrame(animateScroll)
+      requestAnimFrame(animateScroll);
     } else {
       if (callback && typeof (callback) === 'function') {
         // the animation is done so lets callback
-        callback()
+        callback();
       }
     }
-  }
-  animateScroll()
+  };
+  animateScroll();
 }
