@@ -1,25 +1,36 @@
 <template>
-  <el-card style="margin-bottom:20px;">
-    <div slot="header" class="clearfix">
-      <span>关于我</span>
-    </div>
+  <el-card style="margin-bottom: 20px">
+    <template #header>
+      <div class="clearfix">
+        <span>关于我</span>
+      </div>
+    </template>
 
     <div class="user-profile">
       <div class="box-center">
-        <pan-thumb :image="user.avatar" :height="'100px'" :width="'100px'" :hoverable="false">
+        <pan-thumb
+          :image="user.avatar ? user.avatar : ' '"
+          :height="'100px'"
+          :width="'100px'"
+          :hoverable="false"
+        >
           <div>Hello</div>
-          {{ user.role }}
+          {{ user.name }}
         </pan-thumb>
       </div>
       <div class="box-center">
         <div class="user-name text-center">{{ user.name }}</div>
-        <div class="user-role text-center text-muted">{{ user.role | uppercaseFirst }}</div>
+        <div class="user-role text-center text-muted">
+          {{ uppercaseFirst(user.role) }}
+        </div>
       </div>
     </div>
 
-    <div class="user-bio">
+    <!-- <div class="user-bio">
       <div class="user-education user-bio-section">
-        <div class="user-bio-section-header"><svg-icon icon-class="education" /><span>Education</span></div>
+        <div class="user-bio-section-header">
+          <svg-icon icon-class="education" /><span>Education</span>
+        </div>
         <div class="user-bio-section-body">
           <div class="text-muted">
             JS in Computer Science from the University of Technology
@@ -28,7 +39,9 @@
       </div>
 
       <div class="user-skills user-bio-section">
-        <div class="user-bio-section-header"><svg-icon icon-class="skill" /><span>Skills</span></div>
+        <div class="user-bio-section-header">
+          <svg-icon icon-class="skill" /><span>Skills</span>
+        </div>
         <div class="user-bio-section-body">
           <div class="progress-item">
             <span>Vue</span>
@@ -48,29 +61,27 @@
           </div>
         </div>
       </div>
-    </div>
+    </div> -->
   </el-card>
 </template>
 
-<script>
-import PanThumb from '@/components/PanThumb'
-
-export default {
-  components: { PanThumb },
-  props: {
-    user: {
-      type: Object,
-      default: () => {
-        return {
-          name: '',
-          email: '',
-          avatar: '',
-          role: ''
-        }
-      }
-    }
-  }
-}
+<script setup lang="ts">
+import PanThumb from "@/components/PanThumb/index.vue";
+import { defineComponent } from "vue";
+import { uppercaseFirst } from "@/utils/filters";
+const props = defineProps({
+  user: {
+    type: Object,
+    default: () => {
+      return {
+        name: "",
+        email: "",
+        avatar: "",
+        role: "",
+      };
+    },
+  },
+});
 </script>
 
 <style lang="scss" scoped>
