@@ -1,7 +1,7 @@
 package controller
 
 import (
-	httpz2 "fastgin/common/httpz"
+	"fastgin/common/httpz"
 	"fastgin/config"
 	"fastgin/modules/sys/dto"
 	"fastgin/modules/sys/model"
@@ -30,17 +30,17 @@ func NewMenuController() *MenuController {
 // @Accept json
 // @Produce json
 // @Param Authorization header string true "Bearer token"
-// @Success 200 {object} util.ResponseBody
-// @Failure 400 {object} util.ResponseBody
+// @Success 200 {object} httpz.ResponseBody
+// @Failure 400 {object} httpz.ResponseBody
 // @Router /api/auth/menus [get]
 func (mc *MenuController) List(c *gin.Context) {
 	menus, err := mc.menuService.GetMenus()
 	if err != nil {
-		httpz2.ServerError(c, "获取菜单列表失败: "+err.Error())
+		httpz.ServerError(c, "获取菜单列表失败: "+err.Error())
 		return
 	}
 	//util.Success(c, gin.H{"menus": menus}, "获取菜单列表成功")
-	httpz2.Success(c, menus)
+	httpz.Success(c, menus)
 }
 
 // GetMenuTree retrieves the menu tree
@@ -50,17 +50,17 @@ func (mc *MenuController) List(c *gin.Context) {
 // @Accept json
 // @Produce json
 // @Param Authorization header string true "Bearer token"
-// @Success 200 {object} util.ResponseBody
-// @Failure 400 {object} util.ResponseBody
+// @Success 200 {object} httpz.ResponseBody
+// @Failure 400 {object} httpz.ResponseBody
 // @Router /api/auth/menu/tree [get]
 func (mc *MenuController) GetMenuTree(c *gin.Context) {
 	menuTree, err := mc.menuService.GetMenuTree()
 	if err != nil {
-		httpz2.ServerError(c, "获取菜单树失败: "+err.Error())
+		httpz.ServerError(c, "获取菜单树失败: "+err.Error())
 		return
 	}
 	//util.Success(c, gin.H{"MenuTree": menuTree}, "获取菜单树成功")
-	httpz2.Success(c, menuTree)
+	httpz.Success(c, menuTree)
 }
 
 // Create creates a new menu
@@ -71,23 +71,23 @@ func (mc *MenuController) GetMenuTree(c *gin.Context) {
 // @Produce json
 // @Param Authorization header string true "Bearer token"
 // @Param menu body dto.CreateMenuRequest true "Create menu request"
-// @Success 200 {object} util.ResponseBody
-// @Failure 400 {object} util.ResponseBody
+// @Success 200 {object} httpz.ResponseBody
+// @Failure 400 {object} httpz.ResponseBody
 // @Router /api/auth/menu/index [post]
 func (mc *MenuController) Create(c *gin.Context) {
 	var req dto.CreateMenuRequest
 	if err := c.ShouldBind(&req); err != nil {
-		httpz2.BadRequest(c, err.Error())
+		httpz.BadRequest(c, err.Error())
 		return
 	}
 	if err := config.Validate.Struct(&req); err != nil {
 		errStr := err.(validator.ValidationErrors)[0].Translate(config.Trans)
-		httpz2.BadRequest(c, errStr)
+		httpz.BadRequest(c, errStr)
 		return
 	}
 	ctxUser, err := mc.userService.GetCurrentUser(c)
 	if err != nil {
-		httpz2.ServerError(c, "获取当前用户信息失败")
+		httpz.ServerError(c, "获取当前用户信息失败")
 		return
 	}
 	menu := model.Menu{
@@ -110,15 +110,15 @@ func (mc *MenuController) Create(c *gin.Context) {
 	e := copier.Copy(&menu, &req)
 
 	if e != nil {
-		httpz2.ServerError(c, "创建菜单失败: "+e.Error())
+		httpz.ServerError(c, "创建菜单失败: "+e.Error())
 		return
 	}
 	err = mc.menuService.CreateMenu(&menu)
 	if err != nil {
-		httpz2.ServerError(c, "创建菜单失败: "+err.Error())
+		httpz.ServerError(c, "创建菜单失败: "+err.Error())
 		return
 	}
-	httpz2.Success(c, nil)
+	httpz.Success(c, nil)
 }
 
 // Update updates an existing menu by Id
@@ -130,29 +130,29 @@ func (mc *MenuController) Create(c *gin.Context) {
 // @Param Authorization header string true "Bearer token"
 // @Param menuId path int true "Menu Id"
 // @Param menu body dto.CreateMenuRequest true "Update menu request"
-// @Success 200 {object} util.ResponseBody
-// @Failure 400 {object} util.ResponseBody
+// @Success 200 {object} httpz.ResponseBody
+// @Failure 400 {object} httpz.ResponseBody
 // @Router /api/auth/menu/index/{menuId} [put]
 func (mc *MenuController) Update(c *gin.Context) {
 	var req dto.CreateMenuRequest
 	if err := c.ShouldBind(&req); err != nil {
-		httpz2.BadRequest(c, err.Error())
+		httpz.BadRequest(c, err.Error())
 		return
 	}
 	if err := config.Validate.Struct(&req); err != nil {
 		errStr := err.(validator.ValidationErrors)[0].Translate(config.Trans)
-		httpz2.BadRequest(c, errStr)
+		httpz.BadRequest(c, errStr)
 		return
 	}
 	menuId, _ := strconv.Atoi(c.Param("menuId"))
 	if menuId <= 0 {
-		httpz2.BadRequest(c, "菜单ID不正确")
+		httpz.BadRequest(c, "菜单ID不正确")
 		return
 	}
 	//ur := service.NewUserService()
 	ctxUser, err := mc.userService.GetCurrentUser(c)
 	if err != nil {
-		httpz2.ServerError(c, "获取当前用户信息失败")
+		httpz.ServerError(c, "获取当前用户信息失败")
 		return
 	}
 	menu := model.Menu{
@@ -174,16 +174,16 @@ func (mc *MenuController) Update(c *gin.Context) {
 	}
 	e := copier.Copy(&menu, &req)
 	if e != nil {
-		httpz2.ServerError(c, "更新菜单失败: "+e.Error())
+		httpz.ServerError(c, "更新菜单失败: "+e.Error())
 		return
 	}
 	menu.Id = uint(menuId)
 	err = mc.menuService.UpdateMenuById(&menu)
 	if err != nil {
-		httpz2.ServerError(c, "更新菜单失败: "+err.Error())
+		httpz.ServerError(c, "更新菜单失败: "+err.Error())
 		return
 	}
-	httpz2.Success(c, nil)
+	httpz.Success(c, nil)
 }
 
 // BatchDeleteByIds deletes multiple menus by their Ids
@@ -193,22 +193,22 @@ func (mc *MenuController) Update(c *gin.Context) {
 // @Accept json
 // @Produce json
 // @Param Authorization header string true "Bearer token"
-// @Param menuIds body dto.IdListRequest true "BatchDelete menu request"
-// @Success 200 {object} util.ResponseBody
-// @Failure 400 {object} util.ResponseBody
+// @Param menuIds body httpz.IdListRequest true "BatchDelete menu request"
+// @Success 200 {object} httpz.ResponseBody
+// @Failure 400 {object} httpz.ResponseBody
 // @Router /api/auth/menu/index [delete]
 func (mc *MenuController) BatchDeleteByIds(c *gin.Context) {
-	var req httpz2.IdListRequest
+	var req httpz.IdListRequest
 	if err := c.ShouldBind(&req); err != nil {
-		httpz2.BadRequest(c, err.Error())
+		httpz.BadRequest(c, err.Error())
 		return
 	}
 	err := mc.menuService.BatchDeleteMenuByIds(req.Ids)
 	if err != nil {
-		httpz2.ServerError(c, "删除菜单失败: "+err.Error())
+		httpz.ServerError(c, "删除菜单失败: "+err.Error())
 		return
 	}
-	httpz2.Success(c, nil)
+	httpz.Success(c, nil)
 }
 
 // GetUserMenusByUserId retrieves the accessible menus for a user by user Id
@@ -219,22 +219,22 @@ func (mc *MenuController) BatchDeleteByIds(c *gin.Context) {
 // @Produce json
 // @Param Authorization header string true "Bearer token"
 // @Param userId path int true "User Id"
-// @Success 200 {object} util.ResponseBody
-// @Failure 400 {object} util.ResponseBody
+// @Success 200 {object} httpz.ResponseBody
+// @Failure 400 {object} httpz.ResponseBody
 // @Router /api/auth/menu/user/{userId} [get]
 func (mc *MenuController) GetUserMenusByUserId(c *gin.Context) {
 	userId, _ := strconv.Atoi(c.Param("userId"))
 	if userId <= 0 {
-		httpz2.BadRequest(c, "用户ID不正确")
+		httpz.BadRequest(c, "用户ID不正确")
 		return
 	}
 	menus, err := mc.menuService.GetUserMenusByUserId(uint(userId))
 	if err != nil {
-		httpz2.ServerError(c, "获取用户的可访问菜单列表失败: "+err.Error())
+		httpz.ServerError(c, "获取用户的可访问菜单列表失败: "+err.Error())
 		return
 	}
 	//util.Success(c, gin.H{"Menus": menus}, "获取用户的可访问菜单列表成功")
-	httpz2.Success(c, menus)
+	httpz.Success(c, menus)
 }
 
 // GetUserMenuTreeByUserId retrieves the accessible menu tree for a user by user Id
@@ -245,20 +245,20 @@ func (mc *MenuController) GetUserMenusByUserId(c *gin.Context) {
 // @Produce json
 // @Param Authorization header string true "Bearer token"
 // @Param userId path int true "User Id"
-// @Success 200 {object} util.ResponseBody
-// @Failure 400 {object} util.ResponseBody
+// @Success 200 {object} httpz.ResponseBody
+// @Failure 400 {object} httpz.ResponseBody
 // @Router /api/auth/menu/user_tree/{userId} [get]
 func (mc *MenuController) GetUserMenuTreeByUserId(c *gin.Context) {
 	userId, _ := strconv.Atoi(c.Param("userId"))
 	if userId <= 0 {
-		httpz2.BadRequest(c, "用户ID不正确")
+		httpz.BadRequest(c, "用户ID不正确")
 		return
 	}
 	menuTree, err := mc.menuService.GetUserMenuTreeByUserId(uint(userId))
 	if err != nil {
-		httpz2.ServerError(c, "获取用户的可访问菜单树失败: "+err.Error())
+		httpz.ServerError(c, "获取用户的可访问菜单树失败: "+err.Error())
 		return
 	}
 	//util.Success(c, gin.H{"MenuTree": menuTree}, "获取用户的可访问菜单树成功")
-	httpz2.Success(c, menuTree)
+	httpz.Success(c, menuTree)
 }
