@@ -1,8 +1,8 @@
 package controller
 
 import (
-	httpz2 "fastgin/common/httpz"
-	util2 "fastgin/common/util"
+	"fastgin/common/httpz"
+	"fastgin/common/util"
 	"fastgin/config"
 	"fastgin/modules/sys/dto"
 	"fastgin/modules/sys/model"
@@ -31,16 +31,16 @@ func NewUserController() *UserController {
 // @Accept json
 // @Produce json
 // @Param Authorization header string true "Bearer token"
-// @Success 200 {object} util.ResponseBody
-// @Failure 400 {object} util.ResponseBody
+// @Success 200 {object} httpz.ResponseBody
+// @Failure 400 {object} httpz.ResponseBody
 // @Router /api/auth/user/info [get]
 func (uc *UserController) GetUserInfo(c *gin.Context) {
 	user, err := uc.userService.GetCurrentUser(c)
 	if err != nil {
-		httpz2.ServerError(c, "获取当前用户信息失败: "+err.Error())
+		httpz.ServerError(c, "获取当前用户信息失败: "+err.Error())
 		return
 	}
-	userMap, e := util2.StructToMap(user, true, "Id", "UserName", "Mobile", "Avatar", "NickName", "Introduction", "Roles")
+	userMap, e := util.StructToMap(user, true, "Id", "UserName", "Mobile", "Avatar", "NickName", "Introduction", "Roles")
 	if e != nil {
 		fmt.Println(e.Error())
 	}
@@ -48,7 +48,7 @@ func (uc *UserController) GetUserInfo(c *gin.Context) {
 	//fmt.Println(util.Struct2Json(user))
 	//userInfoDto := dto.UserInfoDto{}
 	//copier.Copy(&userInfoDto, &user)
-	httpz2.Success(c, userMap)
+	httpz.Success(c, userMap)
 }
 
 // 获取用户列表
@@ -59,21 +59,21 @@ func (uc *UserController) GetUserInfo(c *gin.Context) {
 // @Produce json
 // @Param Authorization header string true "Bearer token"
 // @Param UserListRequest body dto.UserListRequest true "User list request"
-// @Success 200 {object} util.ResponseBody
-// @Failure 400 {object} util.ResponseBody
+// @Success 200 {object} httpz.ResponseBody
+// @Failure 400 {object} httpz.ResponseBody
 // @Router /api/auth/user/index [get]
 func (uc *UserController) GetUsers(c *gin.Context) {
-	params, e := httpz2.GetFormData(c)
+	params, e := httpz.GetFormData(c)
 	if e != nil {
-		httpz2.BadRequest(c, e.Error())
+		httpz.BadRequest(c, e.Error())
 		return
 	}
-	data, total, err := uc.userService.GetUsersWithRoleIds(httpz2.NewSearchRequest(params))
+	data, total, err := uc.userService.GetUsersWithRoleIds(httpz.NewSearchRequest(params))
 	if err != nil {
-		httpz2.ServerError(c, "获取角色列表失败: "+err.Error())
+		httpz.ServerError(c, "获取角色列表失败: "+err.Error())
 		return
 	}
-	httpz2.Success(c, gin.H{"Data": data, "Total": total})
+	httpz.Success(c, gin.H{"Data": data, "Total": total})
 
 	//
 	//
@@ -108,21 +108,21 @@ func (uc *UserController) GetUsers(c *gin.Context) {
 // @Produce json
 // @Param Authorization header string true "Bearer token"
 // @Param ChangePwdRequest body dto.ChangePwdRequest true "Change password request"
-// @Success 200 {object} util.ResponseBody
-// @Failure 400 {object} util.ResponseBody
+// @Success 200 {object} httpz.ResponseBody
+// @Failure 400 {object} httpz.ResponseBody
 // @Router /api/auth/user/changePwd [post]
 func (uc *UserController) ChangePwd(c *gin.Context) {
 	var req dto.ChangePwdRequest
 
 	// 参数绑定
 	if err := c.ShouldBind(&req); err != nil {
-		httpz2.BadRequest(c, err.Error())
+		httpz.BadRequest(c, err.Error())
 		return
 	}
 	// 参数校验
 	if err := config.Validate.Struct(&req); err != nil {
 		errStr := err.(validator.ValidationErrors)[0].Translate(config.Trans)
-		httpz2.BadRequest(c, errStr)
+		httpz.BadRequest(c, errStr)
 		return
 	}
 
@@ -144,24 +144,24 @@ func (uc *UserController) ChangePwd(c *gin.Context) {
 	// 获取当前用户
 	user, err := uc.userService.GetCurrentUser(c)
 	if err != nil {
-		httpz2.ServerError(c, err.Error())
+		httpz.ServerError(c, err.Error())
 		return
 	}
 	// 获取用户的真实正确密码
 	correctPasswd := user.Password
 	// 判断前端请求的密码是否等于真实密码
-	err = util2.ComparePasswd(correctPasswd, req.OldPassword)
+	err = util.ComparePasswd(correctPasswd, req.OldPassword)
 	if err != nil {
-		httpz2.ServerError(c, "原密码有误")
+		httpz.ServerError(c, "原密码有误")
 		return
 	}
 	// 更新密码
-	err = uc.userService.ChangePwd(user.UserName, util2.GenPasswd(req.NewPassword))
+	err = uc.userService.ChangePwd(user.UserName, util.GenPasswd(req.NewPassword))
 	if err != nil {
-		httpz2.ServerError(c, "更新密码失败: "+err.Error())
+		httpz.ServerError(c, "更新密码失败: "+err.Error())
 		return
 	}
-	httpz2.Success(c, nil)
+	httpz.Success(c, nil)
 }
 
 // 创建用户
@@ -172,20 +172,20 @@ func (uc *UserController) ChangePwd(c *gin.Context) {
 // @Produce json
 // @Param Authorization header string true "Bearer token"
 // @Param CreateUserRequest body dto.CreateUserRequest true "Create user request"
-// @Success 200 {object} util.ResponseBody
-// @Failure 400 {object} util.ResponseBody
+// @Success 200 {object} httpz.ResponseBody
+// @Failure 400 {object} httpz.ResponseBody
 // @Router /api/auth/user/index [post]
 func (uc *UserController) CreateUser(c *gin.Context) {
 	var req dto.CreateUserRequest
 	// 参数绑定
 	if err := c.ShouldBind(&req); err != nil {
-		httpz2.BadRequest(c, err.Error())
+		httpz.BadRequest(c, err.Error())
 		return
 	}
 	// 参数校验
 	if err := config.Validate.Struct(&req); err != nil {
 		errStr := err.(validator.ValidationErrors)[0].Translate(config.Trans)
-		httpz2.BadRequest(c, errStr)
+		httpz.BadRequest(c, errStr)
 		return
 	}
 
@@ -207,7 +207,7 @@ func (uc *UserController) CreateUser(c *gin.Context) {
 	// 当前用户角色排序最小值（最高等级角色）以及当前用户
 	currentRoleSortMin, ctxUser, err := uc.userService.GetCurrentUserMinRoleSort(c)
 	if err != nil {
-		httpz2.ServerError(c, err.Error())
+		httpz.ServerError(c, err.Error())
 		return
 	}
 
@@ -217,11 +217,11 @@ func (uc *UserController) CreateUser(c *gin.Context) {
 	rr := service.NewRoleService()
 	roles, err := rr.GetRolesByIds(reqRoleIds)
 	if err != nil {
-		httpz2.ServerError(c, "根据角色ID获取角色信息失败: "+err.Error())
+		httpz.ServerError(c, "根据角色ID获取角色信息失败: "+err.Error())
 		return
 	}
 	if len(roles) == 0 {
-		httpz2.ServerError(c, "未获取到角色信息")
+		httpz.ServerError(c, "未获取到角色信息")
 		return
 	}
 	var reqRoleSorts []int
@@ -234,7 +234,7 @@ func (uc *UserController) CreateUser(c *gin.Context) {
 
 	// 当前用户的角色排序最小值 需要小于 前端传来的角色排序最小值（用户不能创建比自己等级高的或者相同等级的用户）
 	if currentRoleSortMin >= reqRoleSortMin {
-		httpz2.ServerError(c, "用户不能创建比自己等级高的或者相同等级的用户")
+		httpz.ServerError(c, "用户不能创建比自己等级高的或者相同等级的用户")
 		return
 	}
 
@@ -244,7 +244,7 @@ func (uc *UserController) CreateUser(c *gin.Context) {
 	}
 	user := model.User{
 		UserName:     req.Username,
-		Password:     util2.GenPasswd(req.Password),
+		Password:     util.GenPasswd(req.Password),
 		Mobile:       req.Mobile,
 		Avatar:       req.Avatar,
 		NickName:     &req.NickName,
@@ -256,10 +256,10 @@ func (uc *UserController) CreateUser(c *gin.Context) {
 
 	err = uc.userService.CreateUser(&user)
 	if err != nil {
-		httpz2.ServerError(c, "创建用户失败: "+err.Error())
+		httpz.ServerError(c, "创建用户失败: "+err.Error())
 		return
 	}
-	httpz2.Success(c, nil)
+	httpz.Success(c, nil)
 
 }
 
@@ -272,41 +272,41 @@ func (uc *UserController) CreateUser(c *gin.Context) {
 // @Param Authorization header string true "Bearer token"
 // @Param userId path int true "User Id"
 // @Param CreateUserRequest body dto.CreateUserRequest true "Update user request"
-// @Success 200 {object} util.ResponseBody
-// @Failure 400 {object} util.ResponseBody
+// @Success 200 {object} httpz.ResponseBody
+// @Failure 400 {object} httpz.ResponseBody
 // @Router /api/auth/user/index/{userId} [put]
 func (uc *UserController) Update(c *gin.Context) {
 	var req dto.CreateUserRequest
 	// 参数绑定
 	if err := c.ShouldBind(&req); err != nil {
-		httpz2.BadRequest(c, err.Error())
+		httpz.BadRequest(c, err.Error())
 		return
 	}
 	// 参数校验
 	if err := config.Validate.Struct(&req); err != nil {
 		errStr := err.(validator.ValidationErrors)[0].Translate(config.Trans)
-		httpz2.BadRequest(c, errStr)
+		httpz.BadRequest(c, errStr)
 		return
 	}
 
 	//获取path中的userId
 	userId, _ := strconv.Atoi(c.Param("userId"))
 	if userId <= 0 {
-		httpz2.BadRequest(c, "用户ID不正确")
+		httpz.BadRequest(c, "用户ID不正确")
 		return
 	}
 
 	// 根据path中的userId获取用户信息
 	oldUser, err := uc.userService.GetUserById(uint(userId))
 	if err != nil {
-		httpz2.ServerError(c, "获取需要更新的用户信息失败: "+err.Error())
+		httpz.ServerError(c, "获取需要更新的用户信息失败: "+err.Error())
 		return
 	}
 
 	// 获取当前用户
 	ctxUser, err := uc.userService.GetCurrentUser(c)
 	if err != nil {
-		httpz2.ServerError(c, err.Error())
+		httpz.ServerError(c, err.Error())
 		return
 	}
 	// 获取当前用户的所有角色
@@ -328,11 +328,11 @@ func (uc *UserController) Update(c *gin.Context) {
 	rr := service.NewRoleService()
 	roles, err := rr.GetRolesByIds(reqRoleIds)
 	if err != nil {
-		httpz2.ServerError(c, "根据角色ID获取角色信息失败: "+err.Error())
+		httpz.ServerError(c, "根据角色ID获取角色信息失败: "+err.Error())
 		return
 	}
 	if len(roles) == 0 {
-		httpz2.ServerError(c, "未获取到角色信息")
+		httpz.ServerError(c, "未获取到角色信息")
 		return
 	}
 	var reqRoleSorts []int
@@ -359,19 +359,19 @@ func (uc *UserController) Update(c *gin.Context) {
 		// 如果是更新自己
 		// 不能禁用自己
 		if req.Status == 2 {
-			httpz2.ServerError(c, "不能禁用自己")
+			httpz.ServerError(c, "不能禁用自己")
 			return
 		}
 		// 不能更改自己的角色
 		reqDiff, currentDiff := funk.Difference(req.RoleIds, currentRoleIds)
 		if len(reqDiff.([]uint)) > 0 || len(currentDiff.([]uint)) > 0 {
-			httpz2.ServerError(c, "不能更改自己的角色")
+			httpz.ServerError(c, "不能更改自己的角色")
 			return
 		}
 
 		// 不能更新自己的密码，只能在个人中心更新
 		if req.Password != "" {
-			httpz2.ServerError(c, "请到个人中心更新自身密码")
+			httpz.ServerError(c, "请到个人中心更新自身密码")
 			return
 		}
 
@@ -384,17 +384,17 @@ func (uc *UserController) Update(c *gin.Context) {
 		// 根据path中的userIdID获取用户角色排序最小值
 		minRoleSorts, err := uc.userService.GetUserMinRoleSortsByIds([]uint{uint(userId)})
 		if err != nil || len(minRoleSorts) == 0 {
-			httpz2.ServerError(c, "根据用户ID获取用户角色排序最小值失败")
+			httpz.ServerError(c, "根据用户ID获取用户角色排序最小值失败")
 			return
 		}
 		if currentRoleSortMin >= minRoleSorts[0] {
-			httpz2.ServerError(c, "用户不能更新比自己角色等级高的或者相同等级的用户")
+			httpz.ServerError(c, "用户不能更新比自己角色等级高的或者相同等级的用户")
 			return
 		}
 
 		// 用户不能把别的用户角色等级更新得比自己高或相等
 		if currentRoleSortMin >= reqRoleSortMin {
-			httpz2.ServerError(c, "用户不能把别的用户角色等级更新得比自己高或相等")
+			httpz.ServerError(c, "用户不能把别的用户角色等级更新得比自己高或相等")
 			return
 		}
 
@@ -409,17 +409,17 @@ func (uc *UserController) Update(c *gin.Context) {
 		//	req.Password = string(decodeData)
 		//	user.Password = util.GenPasswd(req.Password)
 		//}
-		user.Password = util2.GenPasswd(req.Password)
+		user.Password = util.GenPasswd(req.Password)
 
 	}
 
 	// 更新用户
 	err = uc.userService.UpdateUser(&user)
 	if err != nil {
-		httpz2.ServerError(c, "更新用户失败: "+err.Error())
+		httpz.ServerError(c, "更新用户失败: "+err.Error())
 		return
 	}
-	httpz2.Success(c, nil)
+	httpz.Success(c, nil)
 
 }
 
@@ -430,21 +430,21 @@ func (uc *UserController) Update(c *gin.Context) {
 // @Accept json
 // @Produce json
 // @Param Authorization header string true "Bearer token"
-// @Param DeleteUserRequest body dto.IdListRequest true "BatchDelete user request"
-// @Success 200 {object} util.ResponseBody
-// @Failure 400 {object} util.ResponseBody
+// @Param DeleteUserRequest body httpz.IdListRequest true "BatchDelete user request"
+// @Success 200 {object} httpz.ResponseBody
+// @Failure 400 {object} httpz.ResponseBody
 // @Router /api/auth/user/index [delete]
 func (uc *UserController) BatchDeleteUserByIds(c *gin.Context) {
-	var req httpz2.IdListRequest
+	var req httpz.IdListRequest
 	// 参数绑定
 	if err := c.ShouldBind(&req); err != nil {
-		httpz2.BadRequest(c, err.Error())
+		httpz.BadRequest(c, err.Error())
 		return
 	}
 	// 参数校验
 	if err := config.Validate.Struct(&req); err != nil {
 		errStr := err.(validator.ValidationErrors)[0].Translate(config.Trans)
-		httpz2.BadRequest(c, errStr)
+		httpz.BadRequest(c, errStr)
 		return
 	}
 
@@ -453,38 +453,38 @@ func (uc *UserController) BatchDeleteUserByIds(c *gin.Context) {
 	// 根据用户ID获取用户角色排序最小值
 	roleMinSortList, err := uc.userService.GetUserMinRoleSortsByIds(reqUserIds)
 	if err != nil || len(roleMinSortList) == 0 {
-		httpz2.ServerError(c, "根据用户ID获取用户角色排序最小值失败")
+		httpz.ServerError(c, "根据用户ID获取用户角色排序最小值失败")
 		return
 	}
 
 	// 当前用户角色排序最小值（最高等级角色）以及当前用户
 	minSort, ctxUser, err := uc.userService.GetCurrentUserMinRoleSort(c)
 	if err != nil {
-		httpz2.ServerError(c, err.Error())
+		httpz.ServerError(c, err.Error())
 		return
 	}
 	currentRoleSortMin := int(minSort)
 
 	// 不能删除自己
 	if slices.Contains(reqUserIds, ctxUser.Id) {
-		httpz2.ServerError(c, "用户不能删除自己")
+		httpz.ServerError(c, "用户不能删除自己")
 		return
 	}
 
 	// 不能删除比自己角色排序低(等级高)的用户
 	for _, sort := range roleMinSortList {
 		if currentRoleSortMin >= sort {
-			httpz2.ServerError(c, "用户不能删除比自己角色等级高的用户")
+			httpz.ServerError(c, "用户不能删除比自己角色等级高的用户")
 			return
 		}
 	}
 
 	err = uc.userService.BatchDeleteUserByIds(reqUserIds)
 	if err != nil {
-		httpz2.ServerError(c, "删除用户失败: "+err.Error())
+		httpz.ServerError(c, "删除用户失败: "+err.Error())
 		return
 	}
 
-	httpz2.Success(c, nil)
+	httpz.Success(c, nil)
 
 }
